validation: centralize axis and shape checks
